update test and move component NoPermissionMessage to components folders
